Moved credentials state to Welcome.js
