+++ src/Components/auth/LoginForm.js
@@ -1,29 +1,22 @@
-import React, {useState} from 'react';
+import React from 'react';
 import dataManager from '../../modules/dataManager'
 
 const LoginForm = props => {
-    const [credentials, setCredentials] = useState({username:"", password: ""})
-
-    const handleSettingCredentials = event => {
-        const stateToChange = {...credentials};
-        stateToChange[event.target.id] = event.target.value;
-        setCredentials(stateToChange);
-    };
 
     const handleLogin = event => {
         event.preventDefault();
 
-        if (credentials.username === "" || credentials.password === "") {
+        if (props.credentials.username === "" || props.credentials.password === "") {
             window.alert('Please complete all fields');
         } else {
-            dataManager.getUsername(credentials.username)
+            dataManager.getUsername(props.credentials.username)
                 .then(userArr => {
                     if (userArr.length < 0) {
                         window.alert('Username does not exist')
-                    } else if (credentials.password !== userArr[0].password) {
+                    } else if (props.credentials.password !== userArr[0].password) {
                         window.alert('Incorrect password')
                     } else {
-                        props.setUser(credentials)
+                        props.setUser(props.credentials)
                     }
                 })
         }
@@ -33,10 +26,10 @@ return (
     <>
         <form>
             <label htmlFor='username'>Username </label>
-            <input onChange={handleSettingCredentials} type='text' id='username' />
+            <input onChange={props.handleSettingCredentials} type='text' id='username' />
 
             <label htmlFor='password'>Password </label>
-            <input onChange={handleSettingCredentials} type='text' id='password' />
+            <input onChange={props.handleSettingCredentials} type='text' id='password' />
 
             <button type='submit' id='sign-up' onClick={handleLogin}>Log in</button>
             <button id='cancel' onClick={props.toggleModal}>Cancel</button>
