+++ src/Components/auth/Welcome.js
@@ -1,10 +1,18 @@
-import React from 'react'
+import React, {useState} from 'react'
 import {createPortal} from 'react-dom';
 import SignUpForm from './SignUpForm'
 import LoginForm from './LoginForm'
 
 const Welcome = props => {
     
+    const [credentials, setCredentials] = useState({username:"", password: ""})
+        
+    const handleSettingCredentials = event => {
+        const stateToChange = {...credentials};
+        stateToChange[event.target.id] = event.target.value;
+        setCredentials(stateToChange);
+    };
+
     const toggleSignUpForm = () => {
         props.showSignUpForm ? props.setShowSignUpForm(false) : props.setShowSignUpForm(true)
     }
@@ -19,7 +27,9 @@ const Welcome = props => {
         <>
             <h1>Re-view</h1>
             {props.showSignUpForm
-            ? createPortal(<SignUpForm 
+            ? createPortal(<SignUpForm
+                                credentials={credentials}
+                                handleSettingCredentials={handleSettingCredentials}
                                 toggleSignUpForm={toggleSignUpForm} 
                                 setUser={props.setUser} 
                                 hasUser={props.hasUser}
@@ -28,7 +38,9 @@ const Welcome = props => {
             }
             <button type="button" onClick={toggleSignUpForm}>Sign up</button>
             {props.showLoginForm
-            ? createPortal(<LoginForm 
+            ? createPortal(<LoginForm
+                                credentials={credentials}
+                                handleSettingCredentials={handleSettingCredentials}
                                 toggleLoginForm={toggleLoginForm} 
                                 setUser={props.setUser} 
                                 hasUser={props.hasUser}
